get model and make prediction, error handling

+++ predict.py
@@ -7,6 +7,7 @@ import numpy as np
 import pickle
 import csv
 import boto3
+import traceback
 from os import walk
 import shutil
 from tensorflow.keras.layers import Conv2D, MaxPool2D, Flatten, LSTM
@@ -56,38 +57,45 @@ def build_predictions(audio_dir):
         
     return y_pred, fn_prob
 
-def write_to_csv(current_dir, csv_name):
-    with open(csv_name + '.csv', 'w') as csvfile:
+def write_single_to_csv(filename):
+    with open('predict.csv', 'w') as csvfile:
         fieldnames = ['fname']
         filewriter = csv.DictWriter(csvfile, fieldnames=fieldnames)
         filewriter.writeheader()
-        for (_, _, files) in walk(current_dir):
-            for filename in files:
-                filewriter.writerow({'fname':filename})
-    print("successfully created predict audio csv")
+        filewriter.writerow({'fname':filename})
+    print("successfully added audio to predict csv")
 
 def download_audio_file(fn):
     bucket_name = 'mm-acs-audio-storage'
     s3_file_path= 'asc_data/audio' + '/' + fn + '.wav'
+    s3_data_path = 'asc_data/data.csv'
     save_as = fn + '.wav'
 
     s3 = boto3.client('s3')
     s3.download_file(bucket_name, s3_file_path, save_as)
-    if not os.path.exists(os.path.join('asc_data','audio')):
-        os.makedirs(os.path.join('asc_data','audio'))
-    shutil.move(save_as, os.path.join('asc_data','audio'))
+    s3.download_file(bucket_name,s3_data_path, 'data.csv')
+    if not os.path.isfile(os.path.join('asc_data','audio', save_as)):
+        if not os.path.exists(os.path.join('asc_data','audio')):
+            os.makedirs(os.path.join('asc_data','audio'))
+        shutil.move(save_as, os.path.join('asc_data','audio'))
+
+def get_downloaded_model_data():
+    df = pd.read_csv('data.csv')
+    return float(df['min']), float(df['max'])
 
 
 def build_single_prediction(fn, audio_dir):
-    y_true = []
-    y_pred = []
-    fn_prob = {}
-    config = Config('conv')
-    download_audio_file(fn)
-    model = load_model(config.model_path)
-    print("Extracting Features from Audio" )
-    
     try:
+        y_true = []
+        y_pred = []
+        fn_prob = {}
+        config = Config('conv')
+        download_audio_file(fn)
+        write_single_to_csv(fn)
+        _min, _max = get_downloaded_model_data()
+        model = load_model(config.model_path)
+        print("Extracting Features from Audio with model data min: " )
+
         rate, wav = wavfile.read(os.path.join(audio_dir, fn+'.wav'))
         y_prob = []
         config.step = int(rate/10)
@@ -97,26 +105,30 @@ def build_single_prediction(fn, audio_dir):
             x = mfcc(sample, rate, numcep=config.nfeat, 
                         nfilt = config.nfilt, nfft = config.nfft)
             #normalize x
-            x = (x - config.min)/(config.max - config.min)
+            x = (x - _min)/(_max - _min)
             
             x = x.reshape(1, x.shape[0], x.shape[1], 1)
         
             y_hat = model.predict(x)
             y_prob.append(y_hat)
             y_pred.append(np.argmax(y_hat))
-        
-        fn_prob[fn]= np.mean(y_prob, axis=0).flatten()
 
-        y_probs = get_probabilities(fn_prob)
+        fn_prob[fn]= np.mean(y_prob, axis=0).flatten()
+        
+        y_probs, classes = get_probabilities(fn_prob)
 
         y_pred = [classes[np.argmax(y)] for y in y_probs]
-            
+
+        remove_acs_data_dir()
+
         return y_pred[0]
     except:
-        logging.exception("prediction failed")
+        return traceback.format_exc()
+
+def remove_acs_data_dir():
+    shutil.rmtree('asc_data')
 
 def get_dataframes():
-    write_to_csv('predict', 'predict')
     df = pd.read_csv('predict.csv')
     dfcats = pd.read_csv('cats.csv')
     return df, dfcats
@@ -131,7 +143,7 @@ def get_probabilities(fn_prob):
         y_probs.append(y_prob)
         for c,p in zip(classes, y_prob):
             df.at[i, c] = p
-    return y_probs
+    return y_probs, classes
 
 if __name__ == "__main__":
     #the info given in instruments.csv will have to be derived from directory labels with new data
